+++ src/main/java/org/goormthon/beotkkotthon/rebook/controller/ChallengeController.java
@@ -10,8 +10,8 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.goormthon.beotkkotthon.rebook.annotation.UserId;
 import org.goormthon.beotkkotthon.rebook.dto.common.ResponseDto;
-import org.goormthon.beotkkotthon.rebook.dto.response.ChallengeDto;
-import org.goormthon.beotkkotthon.rebook.dto.response.ChallengeListDto;
+import org.goormthon.beotkkotthon.rebook.dto.response.challenge.ChallengeDto;
+import org.goormthon.beotkkotthon.rebook.dto.response.challenge.ChallengeListDto;
 import org.goormthon.beotkkotthon.rebook.usecase.challenge.ReadChallengeListUseCase;
 import org.goormthon.beotkkotthon.rebook.usecase.challenge.ReadChallengeUseCase;
 import org.springframework.web.bind.annotation.GetMapping;
