+++ src/main/java/org/goormthon/beotkkotthon/rebook/controller/ChallengeRoomController.java
@@ -8,9 +8,8 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.goormthon.beotkkotthon.rebook.dto.common.ResponseDto;
-import org.goormthon.beotkkotthon.rebook.dto.response.ChallengeListDto;
-import org.goormthon.beotkkotthon.rebook.dto.response.ChallengeRoomUserDto;
-import org.goormthon.beotkkotthon.rebook.usecase.challengeRoom.ReadChallengeRoomUserListUseCase;
+import org.goormthon.beotkkotthon.rebook.dto.response.challengeroom.ChallengeRoomUserDto;
+import org.goormthon.beotkkotthon.rebook.usecase.challengeroom.ReadChallengeRoomUserListUseCase;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
