--- src/main/java/org/goormthon/beotkkotthon/rebook/eventcase/MatchingEventCase.java
@@ -1,7 +1,0 @@
-package org.goormthon.beotkkotthon.rebook.eventcase;
-
-import org.goormthon.beotkkotthon.rebook.event.EnterChallengeWaitRoomEvent;
-
-public interface MatchingEventCase {
-    void execute(EnterChallengeWaitRoomEvent event);
-}
