+++ src/main/java/org/goormthon/beotkkotthon/rebook/usecase/challenge/ReadChallengeListUseCase.java
@@ -1,7 +1,7 @@
 package org.goormthon.beotkkotthon.rebook.usecase.challenge;
 
 import org.goormthon.beotkkotthon.rebook.annotation.UseCase;
-import org.goormthon.beotkkotthon.rebook.dto.response.ChallengeListDto;
+import org.goormthon.beotkkotthon.rebook.dto.response.challenge.ChallengeListDto;
 
 import java.util.List;
 import java.util.UUID;
